test(alias): relocate alias-resolution test out of matcher/, share db fixture
The alias-resolution test exercises L2Account resolution, not the matcher, so it does
not belong under tests/unit/matcher/. Move it to tests/unit/alias/ and promote the
db/anyio_backend fixtures to tests/unit/conftest.py so both suites share them.
